api/requirements.txt
fastapi
uvicorn
python-chess
cassandra-driver